Add member name to the queried data
Still need the translation before we can proceed, though.
