+++ apollo/src/main/graphql/com/hedvig/android/owldroid/graphql/home.graphql
@@ -1,4 +1,7 @@
 query HomeQuery {
+  member {
+    firstName
+  }
   contracts {
     switchedFromInsuranceProvider
     status {
