+++ app/src/main/java/com/hedvig/app/feature/home/ui/HomeModel.kt
@@ -1,0 +1,9 @@
+package com.hedvig.app.feature.home.ui
+
+sealed class HomeModel {
+    sealed class BigText : HomeModel() {
+        data class Pending(
+            val name: String
+        ) : BigText()
+    }
+}
