+++ app/src/main/java/com/hedvig/app/util/GenericDiffUtilCallback.kt
@@ -1,0 +1,16 @@
+package com.hedvig.app.util
+
+import androidx.recyclerview.widget.DiffUtil
+
+class GenericDiffUtilCallback<T>(
+    private val old: List<T>,
+    private val new: List<T>
+) : DiffUtil.Callback() {
+    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int) =
+        old[oldItemPosition] == new[newItemPosition]
+
+    override fun getOldListSize() = old.size
+    override fun getNewListSize() = new.size
+    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int) =
+        old[oldItemPosition] == new[newItemPosition]
+}
