+++ testdata/src/main/java/com/hedvig/app/testdata/feature/home/builders/HomeDataBuilder.kt
@@ -3,9 +3,13 @@ package com.hedvig.app.testdata.feature.home.builders
 import com.hedvig.android.owldroid.graphql.HomeQuery
 
 data class HomeDataBuilder(
-    private val contracts: List<Status> = emptyList()
+    private val contracts: List<Status> = emptyList(),
+    private val firstName: String = "Test"
 ) {
     fun build() = HomeQuery.Data(
+        member = HomeQuery.Member(
+            firstName = firstName
+        ),
         contracts = contracts.map { c ->
             HomeQuery.Contract(
                 switchedFromInsuranceProvider = null,
